requirements.txt
astroid==1.6.2
autopep8==1.3.4
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.3.1
pylint==1.8.3
python-pushover==0.3
requests==2.18.4
rope==0.10.7
six==1.11.0
urllib3==1.22
wrapt==1.10.11
